added new compoent for requested changes display on the releases tab
